ti_config_fragments: rpmsg: Add Keystone remoteproc module
Add the Keystone remoteproc module to the rpmsg Kconfig fragment.
This is the minimum required module for enabling remoteproc/rpmsg
communication with the DSP remote processors on various Keystone 2
family of SoCs. The other dependent virtio/rpmsg modules are
automatically enabled when using multi_v7_defconfig as the base
defconfig.

Signed-off-by: Suman Anna <s-anna@ti.com>