HACK: remoteproc: silence devmem error for rprocs with no MMU
The R5F remote processors on K3 AM65x SoCs do not have an MMU, but
the current firmware images still use a TYPE_DEVMEM entry for the
initial 1 MB of the device carveout pool (akin to the usage on the
DRA7xx/AM57xx SoCs). This entry is currently exercised by the firmware
code to perform some address translations, but the current error
check fails to load such firmwares. Silence the error for the
time being to successfully load such firmwares.

Signed-off-by: Suman Anna <s-anna@ti.com>