arm64: dts: ti: k3-am65-main: Add hwspinlock node
The Main NavSS block on AM65x SoCs contains a HwSpinlock IP instance
that is similar to the IP on some OMAP SoCs. Add the DT node for this
on AM65x SoCs. The node is present within the NavSS block, and is
added as a child node under the cbass_main node.

NOTE:
The node should be moved into a navss_main interconnect node if the
local interconnect of Main NavSS is represented as a child node of
the cbass_main interconnect node in the future.

Signed-off-by: Suman Anna <s-anna@ti.com>