remoteproc/k3-r5: add loading support for on-chip SRAM regions
The K3 SoCs has various internal on-chip SRAM memories like the SRAM
within the MCU domain or the shared MSMC RAM within NavSS that can be
used for multiple purposes. One such purpose is to have the R5F cores
to use a portion of such on-chip SRAM for fast-access data or to
directly execute code.

Add support to the K3 R5 remoteproc driver to parse and support
loading into such memories.

These SRAM regions are completely optional as not all firmware images
require these memories, and any such memory has to be reserved as such
in the DTS files.

Signed-off-by: Suman Anna <s-anna@ti.com>