dt-bindings: remoteproc: k3-r5f: Update bindings for on-chip SRAM support
The K3 SoCs have various internal on-chip SRAM memories like the SRAM
within the MCU domain or the shared MSMC RAM within NavSS that can be
used for multiple purposes. One such purpose is to have the MCU domain
R5F cores to use a portion of such on-chip SRAM for executing code.

Update the K3 R5F remoteproc bindings to add the properties required to
support the optional on-chip SRAM loading support.

Signed-off-by: Suman Anna <s-anna@ti.com>