defconfig_builder: Add the defconfig build script and support configs
Add the defconfig builder to the 4.19 kernel.  In addition the defconfig
maps for the v7 and v8 architectures have been added with the basic config
options selected.

These are the base configs to start and may be modified based on additional or
unneeded features.

To see a list of possible defconfigs to build use

defconfig_builder.sh -l

Signed-off-by: Dan Murphy <dmurphy@ti.com>