HACK: drivers: net: cpsw: switch-config: unknown vlan handling
Currently only known VLAN id packets are handled inside the switch,
when an unknown VLAN is received, the switch will drop the packets.
This ioctl allows user to subscribe the ports for unknown VLAN id
packet forwarding with combinations like packet with multicst destination
address present which is present in ALE table with unknown VLAN id and
packet with multicast destination address which is not present in ALE
table with unknown VLAN id, then switch will start forwarding these
packets as per configured by the user. Additional features of the
unknown vlan handling of the switch other than port forwarding is to
untag on egress for external ports.

Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>