net: ethernet: ti: ale: add support for mac-only mode
New CPSW ALE version, available on K3 AM654 SoCs family, allows to switch
any external port to MAC only mode. When MAC only mode enabled this port be
treated like a MAC port for the host. All traffic received is only sent to
the host. The host must direct traffic to this port as the lookup engine
will not send traffic to the ports with the p0_maconly bit set and the
p0_no_learn also set. If p0_maconly bit is set and the p0_no_learn is not
set, the host can send non-directed packets that can be sent to the
destination of a MacOnly port. It is also possible that The host can
broadcast to all ports including MacOnly ports in this mode.

This patch add ALE supprt for MAC only mode.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>