ti_config_fragments: rpmsg: Add RPMsg domain config fragment file
Add the initial IPC/RPMsg domain config fragment file. This will
form the foundation for adding all the RPMsg related module
configurations, to be used by the TI Linux integration kernel.

Also, enable the fully upstreamed (on OMAP2+ SoCs) OMAP HwSpinlock
and OMAP Mailbox drivers including the corresponding HwSpinlock and
Mailbox cores. These are chosen to be built-in as it is the preferred
choice for these drivers.

Signed-off-by: Suman Anna <s-anna@ti.com>