drm/tidss: use drm_fbdev_generic_setup
The driver crashes when calling drm_fbdev_cma_fini(), and the reason
seems to be that drm_dev_unregister() already releases the fbdev, or
parts of it. The exact problem is unclear.

The DRM framework has a new way to setup the fbdev, by calling
drm_fbdev_generic_setup. This is used in the mainline, e.g.
2f690fad237adea1bdce8ed9026cbf43cd9d8c46 ("drm/rcar-du: Use
drm_fbdev_generic_setup()").

This patch does a similar change to tidss, which also seems to get rid
of the crash problem.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>