ARM: dts: am33xx: add device tree entry for SGX
SGX device entry is added to base am33xx dtsi file.

Note: The SGX is disabled by default because it is an optional module
and only some AM335x variants contain this module, such as AM3358 and
AM3357.  The status should be overwritten as "OK" at the corresponding
board.dts

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Eric Ruei <e-ruei1@ti.com>
Signed-off-by: Anand Balagopalakrishnan <anandb@ti.com>
[praneeth@ti.com: rebase from graphics-ti-linux-4.14.y]
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>