net: keystone: add support of keystone common serdes phy driver for gbe
Prior to this patch, gbe inherits the SerDes configurations from
u-boot.  This patch updates the Keystone ethernet driver to use the
generic Keystone serdes phy driver for serdes initialization
and configuration.  The common serdes phy driver supports peripherals
on Keystone platforms that require serdes namely 1GBE, 10GBE, PCIE,
and SRIO. This also replace the netcp_xgbepcsr.c for 10GBE and is
thus removed as part of this patch.

Additionally, this patch introduced serdes lane recovery procedure for
1gbe and 10gbe links.  When the remote device that the 10gbe device is
connected to is restarted for whatever reason, the receive direction of
a 10gbe serdes lane will need to be reset in order to sync up again with
the remote device's transmit pattern. When 10gbe driver detects a serdes
lane down, it will start a periodic attempt to recover the receive
direction of that lane. Similar changes done for 1gbe serdes lane receive
recovery when MAC_MAC_ENFORCED interface configuration is used.

This patch also adopts the remodeling of one serdes lane as a separate
PHY device. Hence phy_init() is invoked per each Ethernet interface.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Rex Chang <rchang@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>