ARM: dts: keystone: enable all dma channels for NetCP
In Keystone, there are user space low level drivers using uio interface
that pushes packets to NetCP queues. So it is necessary to enable all of
the channels associated with these queues. Enable the existing DT
property in pkt dma device bindings to do this and ensure backward
compatibility with existing applications when they move to work with
latest kernel.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>