net: netcp: nss 2u: use syscon phandle to access subsystem
Like other netcp devices, use syscon phandle to access
subsystem area of NSS 2u switch so that the driver can read
the ss version of the switch in a similar way. Also fix the
code not to access the sgmii registers on NSS 2u switch as
it doesn't exist and rgmii h/w interface is used. So change
code to use regmap_read() to read RGMII status register.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>