ti_config_fragments: v8_rpmsg: Enable PRUSS remoteproc support
The PRUSS platform/remoteproc drivers are enhanced to support the newer
PRUSS IPs called ICSSG on the K3 family of SoCs. Add support to build
these PRUSS platform/remoteproc drivers including the remoteproc core
to enable the ICSSG instances. The PRU RPMsg driver and the dependent
virtio/rpmsg modules are also explicitly enabled as these are needed
to showcase the remoteproc/rpmsg communication with the PRU/RTU cores
within the ICSSG instances. All of these drivers will be built as
modules by default.

Signed-off-by: Suman Anna <s-anna@ti.com>