ARM: DRA7: hwmod data: Add MMU data for DSPs
Add the data structures for representing the MMUs within the DSP
processor subsystems present in DRA7xx/AM57xx SoCs. The DRA7xx
family of SoCs usually have one or two DSPs. The DRA74x/DRA76x
family has two DSPs, while DRA72x/DRA71x has only a single DSP.
Each DSP subsystem has two MMUs, one for the processor core and
the other for the internal EDMA block. The hwmod data for the
second DSP is only added for DRA74x/DRA76x family of SoCs.

Both these MMUs share a common reset line, the MMU on the EDMA
port is expected to be mirror-programmed alongside the primary
MMU. The reset data is added to both the MMUs to allow the
omap_hwmod layer to skip the enabling and idling of these devices,
as that would require the reset be released, which is outside
the scope of the hwmod core code. The other PRCM data fields are
also skipped for both the second MMUs, this will be handled as
part of the primary MMU enabling sequence. The second MMUs
needed the HWMOD_OMAP4_NO_CONTEXT_LOSS_BIT as the .context_offs
field does not use a 0 value as a no-op. Without this, CCS throws
the following error upon attempting to connect to DSPs:
  "C66xx_DSP: Error connecting to the target: (Error -6305) PRSC
   module failed to write to a router register."

The pdata quirks will also not be added for the second MMU as the
OMAP IOMMU driver releases the reset once and is expected to program
both the MMUs together.

Signed-off-by: Suman Anna <s-anna@ti.com>