ARM: dts: keystone-k2g-ice: Add PRUSS Ethernet support
The K2G ICE board supports 4 PRU Ethernet ports using DP83822 PHYs and
providing link detection through GPIO interrupts. Each PRUSS instance
supports 2 ports. Add support for this.

The PRUSS Ethernet pins have to be added to the PRUSS nodes so that the
PRUSS PHYs can be bootstrapped properly when the MDIO node probes for the
PHYs.

The PR1_PRU1_GPO1 pin (PADCONFIG_202) cannot be setup from the kernel
due to a silicon bug that corrupts writes to PADCONFIG_202
(and PADCONFIG_194) when BOOT_CFG RSTMUX8 LOCK bit is set.
We expect that this bit is configured correctly in the bootloader
before setting the RSTMUX8 LOCK bit.

Based on initial work by Suman Anna.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>