drm/bridge: sii902x: Disable mclk only if it was successfully enabled
ASoC snd_soc_dai_ops shutdown() is called even if hw_params() was not
successful. This means that the mclk may not be enabled when shutdown()
is called. In such a case the unconditional clk_disable_unprepare()
causes an ugly WARN.

The commit adds a data member to track mclk enabled state. At the same
time the clk_prepare_enable() is moved before clk_get_rate(), because
according to clk documentation clk_get_rate() call is not valid if the
clock source is not enabled.

Signed-off-by: Jyri Sarha <jsarha@ti.com>