arm64: dts: ti: k3-am654-pcie-usb2: Add USB host port
This card has a USB 2.0 dual-role port. As the base board
already supports a 2.0 dual-role port we enable the
port on the SERDES card to be a host only port.

This will prevent user confusion as having 2 ports in
device mode often leads to confusion as to which port
is bound to the gadget function driver.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>